Refactor UnityFile dependencies: they are now transparent for actions
Create an intermediate FileDependency between any action and a UnityFile, with a static dependency from the output file to the unity node.
This solves the issue of GetStaticInputFiles(), which would not list UnityFile depencies since they are BuildGeneratedFile and not BuildSourceFile.
Now we correctly enumerate the file dependency since it does implement BuildSourceFile, simplying unity handling.
This commit also factors this behavior with BuildContext.OutputFile(), which already dealt with this issue.