Fix incorrect ExportIndex in CreateActionRules()
Previous code was appending the export file at the end of output files,
and then assumed an ExportIndex == len(OutputFiles) - 1.
But OutputFiles are then sorted to force determinism, and items can be reordered.
This error was detected because the extension for precompiled object on windows changed from `pch.obj` to simply `.obj`,
and `.obj` is less than `.pch`.