Replace bytes.Buffer by custom AsyncReader/AsyncWriter
Like bytes.Buffer, they use an internal buffer to batch syscalls (which if very effective for small writes).
But unlike bytes.Buffer, they use a page recycler and will defer syscalls to dedicated thread pools.
AsyncReader will read as fast as possible the whole inner reader, dividing it in pages of fixed size ready for calling gorouting thanks to a channel.
AsyncWriter will never block of write, since it always push syscall in a threads pool to run asynchronously.
AsyncTransientIOCopy is now using both of these helpers instead of implementing bespoke parallelisation.
Note we can't more than 2 reading tasks running concurrently in the whole program, but only 1 write is allowed:
this is due to the nature of read, which we will preempt a whole worker until file has finished to read (or when signaled to stop).
Overall this is good speed optimisation for the whole build process.