Refactor EnumSet[T, E] to avoid variables escaping to heap
Relying of pointer to T wrapped in an interface is a bad idea in GO:
do to the limited analysis possible, compiler will always escape to heap pointer references passed through an interface :'(
It lead to EnumSet[T, E].Has() being visible in memory profiler, which is particulary sad for a int32 variable :'(((

The whole was refactored using a type set contraint limiting T to only integral types, allowing us to set *T's value without relying on FromOrd().
Most of the API is not escape-free, except for Set/Test/Select() which still needs interface wrapping to handle parsing values.