Remove comparable constraint from generic Recycler[T]
This solve a linter issue about T "must be pointer-like to avoid allocations".
Comparable was used only by recylcer's debug variant, which relies on a shared map to check sanity of cached objects lifetime.
SharedMap[T] needed a comparable type only to pin the shared map to a vanilla map[]: this unused function was removed, and T can now be `any` everywhere :)