Disable ReaderFrom/WriterTo interfaces for ObservableIO
ObservableIO is used for performance reporting exclusively and not used when not needed,
but we lose all notion of progress when using ReaderFrom/WriterTo interface since only 1 big call will be made.
So, we don't use ReaderFrom/WriterTo interface detection to keep good user-feedback, even if it implies using a less optimal copy interface.
(Note that optimal code path is still used when the program runs headless or on a un-interactive terminal).