Fix pinned log issue which was deleting the last log line when first open
This one tooked some time to be identified: when using a progress scope or a pinned message for the first time,
we were outputing the ASCII escape sequence to delete the last pinned messages (this is how pinned messages work) with 0 line to remove.
The issue was due to this first escape sequence: terminal seems to always delete 1 line, even if we specified 0.
So the fix as to not output this special string when no pinned messages were painted on screen.