Use cut emojis for build summary and worker thread identification
Each thread is assigned a random emoji at start, it is then used as a substitute for previous worker index log annotation (eg [W:03]).
This is really more cut, and used only from an interactive terminal (for which I assume the typical user to have an unicode font setup in 2025).