Refactor the model: ditch -namespace.json, keep only -module.json
All modules are discovered thanks to globing: it simplified many things and also remove a bunch of  code, which was not very useful. Very few interesting usage of passing parameters from the  namespace  were interesting, compared  to the burden of keeping more json files in  check for the user.

This does have one important consequence for the build graph described in the following commit, but this change seems  reasonnable now