Finally got rid of the flickering on progress bars
The issue disapears when not using per character true rgb colors on the whole progress bar.  I think it appears because of the size of the line overflowing some internal buffer of the terminal,  and  not particularly because of rgb colors as the actual fix was to use the  same rgb color for the whole bar.